+++ x-pack/plugins/security_solution/scripts/endpoint/resolver_generator_script.ts
@@ -9,8 +9,6 @@ import { Client, ClientOptions } from '@elastic/elasticsearch';
 import { ResponseError } from '@elastic/elasticsearch/lib/errors';
 import { KbnClient, ToolingLog } from '@kbn/dev-utils';
 import { AxiosResponse } from 'axios';
-import { KibanaConfig } from '@kbn/dev-utils/target/kbn_client/kbn_client_requester';
-import fetch, { RequestInit as FetchRequestInit } from 'node-fetch';
 import { indexHostsAndAlerts } from '../../common/endpoint/index_data';
 import { ANCESTRY_LIMIT } from '../../common/endpoint/generate_data';
 import { FLEET_SETUP_API_ROUTES, SETUP_API_ROUTE } from '../../../ingest_manager/common/constants';
@@ -18,29 +16,7 @@ import {
   CreateFleetSetupResponse,
   PostIngestSetupResponse,
 } from '../../../ingest_manager/common/types/rest_spec';
-
-export class KbnClientWithApiKeySupport extends KbnClient {
-  private kibanaUrlNoAuth: string;
-  constructor(log: ToolingLog, kibanaConfig: KibanaConfig) {
-    super(log, kibanaConfig);
-    const kibanaUrl = this.resolveUrl(kibanaConfig.url);
-    const matches = kibanaUrl.match(/(https?:\/\/)(.*\:.*\@)(.*)/);
-    // strip auth from url
-    this.kibanaUrlNoAuth =
-      matches && matches.length >= 3
-        ? matches[1] + matches[3].replace('/', '')
-        : kibanaUrl.replace('/', '');
-  }
-  /**
-   * The fleet api to enroll and agent requires an api key when you mke the request, however KbnClient currently does not support sending an api key with the request. This function allows you to send an api key with a request.
-   */
-  requestWithApiKey(path: string, init?: RequestInit | undefined): Promise<Response> {
-    return (fetch(
-      `${this.kibanaUrlNoAuth}${path}`,
-      init as FetchRequestInit
-    ) as unknown) as Promise<Response>;
-  }
-}
+import { KbnClientWithApiKeySupport } from './kbn_client_with_api_key_support';
 
 main();
